Added UART support, timeout timer and new build_time string
Wihtin this commit the UART support to transfer strings from the board to an other endpoint (e.g. a PC) is supported. Furthermore a new system build string is added into the build.h file in the format (YYYY-MM-DD-HH-MM-SS) to better identify the firmware-version later.

Please note: There is a known bug: Even if the UART-RX interrupt is disabled at the moment, the PICLCD stop watch is crashing if teraterm is connecting to the COM-PORT.