Added save to eeprom functionality
The slot where the last measurement was saved to will be displayed on the LCD like "-> #0000" (data was storred at EEPROM address 0x0002) or "-> # 0001" (data was storred at EEPROM address 0x0006). Four bytes are neccessarry for every measurement and the first two bytes (address 0x0000 and 0x0001 are holding the address of the next free address).